+++ src/create-card/domain/create-card.domain.js
@@ -3,7 +3,7 @@ const { CreateCardValidation } = require('../schema/input/create-card.input');
 const { createCardService } = require('../service/create-card.service');
 
 const createCardDomain = async (commandPayload, commandMeta) => {
-  const payload = JSON.parse(commandPayload.Message);
+  const payload = JSON.parse(commandPayload.Message); //pasar esto al handler
   new CreateCardValidation(payload, commandMeta);
 
   await createCardService(payload);
