+++ src/create-gift/handler/create-gift.js
@@ -2,11 +2,13 @@ const { batchEventMapper } = require('ebased/handler');
 const inputMode = require("ebased/handler/input/batchEventQueue");
 const outputMode = require("ebased/handler/output/batchEventConfirmation");
 
-const { createGiftDomain } = require('../domain/create-gift.domain');
+const { createGiftDomain } = require('../domain/create-gift.domain');
 
-module.exports.handler = async (events, context) => batchEventMapper(
-  { events, context },
-  inputMode,
-  createGiftDomain,
-  outputMode,
-);
+module.exports.handler = async (events, context) => {
+  return batchEventMapper(
+    { events, context },
+    inputMode,
+    createGiftDomain,
+    outputMode,
+  );
+}
