+++ src/list-client/service/list-client.service.js
@@ -1,6 +1,7 @@
 const dynamodb = require('ebased/service/storage/dynamo');
 
 const listClients = async () => {
+  //TODO: Cambiar scan por query
   const params = {
     TableName: process.env.CLIENTS_TABLE
   }
